rename test_custom_reports.py to test_static_reports.py
