+++ corehq/apps/userreports/tests/__init__.py
@@ -1,7 +1,7 @@
 from .test_app_manager_integration import *
 from .test_columns import *
 from .test_static_data_sources import *
-from .test_custom_reports import *
+from .test_static_reports import *
 from .test_export import *
 from .test_expressions import *
 from .test_filters import *
