+++ corehq/apps/userreports/tests/test_static_reports.py
